Several bug fixes (including stuck USB transfers).
- Fixed a race condition in the USB stack causing the processing thread to not always be woken up by a new event in the IRQ queues.
- Fixed a potential memory corruption bug in the STM32 I2C driver.
- Fixed a bug in FMCSDRAM::SetRefreshRate() causing the refresh cycle to be one clock cycle longer than requested.
- Added buffer size validation to RemoteSignal argument serialization.
- Updated the MPU settings to deny access to any unused address ranges, deny all access to the ITCM RAM at address 0, deny write access to the DRAM copy of the text segment and prevent execution for all other areas except the FLASH and DRAM text segments.
- Fixed some nullptr access bugs (that was masked because there was read/write access to the 64KB ITCM RAM at address 0).
- When giving a Ptr<> pointer a new value, the reference count of the new object is increased before the count of the previous object is decreased. This prevent a potential bug where the destruction of the previous objects makes the reference
- Lots of code cleanup including improved const correctness, more proper integer types used, replacing numerical literals with named consts.

+++ Include/GUI/Region.h
@@ -68,8 +68,8 @@ public:
     std::vector<IRect> m_Rects;
 
 private:
-    int FindUnusedClip(int prevUnused, int lastToCheck);
-    int AddOrReuseClip(int prevUnused, int lastToCheck, const IRect& frame);
+    size_t FindUnusedClip(size_t prevUnused, size_t lastToCheck);
+    size_t AddOrReuseClip(size_t prevUnused, size_t lastToCheck, const IRect& frame);
     void RemoveUnusedClips(size_t firstToCheck, size_t lastToCheck);
   
 };
